Revert "refactor: move s3 config handling to shared module"
This reverts commit ec1ac055cdd9c2e5deb264a9c691c84fd54e9625.
