+++ cdn-server/package.json
@@ -25,7 +25,6 @@
   "dependencies": {
     "@aws-sdk/client-s3": "^3.529.1",
     "@hono/node-server": "1.11.0",
-    "@wundergraph/cosmo-shared": "workspace:*",
     "@wundergraph/cosmo-cdn": "workspace:*",
     "dotenv": "^16.4.5",
     "hono": "4.2.7"
