+++ controlplane/src/types/index.ts
@@ -576,3 +576,11 @@ export interface SchemaLintIssues {
   warnings: LintIssueResult[];
   errors: LintIssueResult[];
 }
+
+export interface S3StorageOptions {
+  url: string;
+  region?: string;
+  endpoint?: string;
+  username?: string;
+  password?: string;
+}
