--- shared/src/types/index.ts
@@ -1,7 +1,0 @@
-export interface S3StorageOptions {
-  url: string;
-  region?: string;
-  endpoint?: string;
-  username?: string;
-  password?: string;
-}
