+++ shared/src/utils/util.ts
@@ -1,6 +1,4 @@
-import * as S3 from '@aws-sdk/client-s3';
 import { SubscriptionProtocol, WebsocketSubprotocol } from '../router-config/builder.js';
-import { S3StorageOptions } from '../types/index.js';
 
 export function delay(t: number) {
   return new Promise((resolve) => setTimeout(resolve, t));
@@ -77,47 +75,4 @@ export function isValidWebsocketSubprotocol(protocol: WebsocketSubprotocol) {
       return false;
     }
   }
-}
-
-export function createS3ClientConfig(bucketName: string, opts: S3StorageOptions): S3.S3ClientConfig {
-  const url = new URL(opts.url);
-  const { region, username, password } = opts;
-  const forcePathStyle = !isVirtualHostStyleUrl(url);
-  const endpoint = opts.endpoint || (forcePathStyle ? url.origin : url.origin.replace(`${bucketName}.`, ''));
-
-  const accessKeyId = url.username || username || '';
-  const secretAccessKey = url.password || password || '';
-
-  if (!accessKeyId || !secretAccessKey) {
-    throw new Error('Missing S3 credentials. Please provide access key ID and secret access key.');
-  }
-
-  if (!region) {
-    throw new Error('Missing region in S3 configuration.');
-  }
-
-  return {
-    region,
-    endpoint,
-    credentials: {
-      accessKeyId,
-      secretAccessKey,
-    },
-    forcePathStyle,
-  };
-}
-
-export function extractS3BucketName(s3Url: string) {
-  const url = new URL(s3Url);
-
-  if (isVirtualHostStyleUrl(url)) {
-    return url.hostname.split('.')[0];
-  }
-
-  // path based style
-  return url.pathname.slice(1);
-}
-
-export function isVirtualHostStyleUrl(url: URL) {
-  return url.hostname.split('.').length > 2;
 }
